updated the home screen and moved popover to ui components
